Remove SugEngine and rename SugEngineImpl to SugEngine
